Switch sorting of rabbitmq comparaison and add variable to defaults
the rabbitmq_required was comparing first with a version number then with 'latest', but 'latest'
can't be compared with a version, so it would fail before going to the 2nd step. Turning this
around solved the issue.
I also added the 'tower_release_version' variable to the defaults.
And while I was at it, I've add Vim swap files to gitignore.